# services/auth-service/storage/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: EMP Auth Service API
  version: 0.1.0
  description: Authentication and authorization service for Enterprise Marketplace platform
  contact:
    name: Engineering Team
    email: engineering@enterprise-marketplace.com
servers:
  - url: http://localhost:8001/api
    description: Local development server
  - url: https://auth.emp-marketplace.com/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a customer account
      description: Create a new customer account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              customer:
                summary: Customer registration
                value:
                  name: John Doe
                  email: john.doe@example.com
                  password: SecurePass123!
                  password_confirmation: SecurePass123!
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      description: Authenticate user credentials and receive access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              valid:
                summary: Valid credentials
                value:
                  email: john.doe@example.com
                  password: SecurePass123!
      responses:
        '200':
          description: Tokens issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - password_confirmation
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        password:
          type: string
          minLength: 8
          format: password
          description: User's password (min 8 characters)
          example: SecurePass123!
        password_confirmation:
          type: string
          format: password
          description: Password confirmation
          example: SecurePass123!
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        password:
          type: string
          format: password
          description: User's password
          example: SecurePass123!
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            access_token:
              type: string
              description: JWT access token
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refresh_token:
              type: string
              description: JWT refresh token
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              description: Access token expiration time in seconds
              example: 900
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        uuid:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: 2026-04-20T12:00:00Z
        created_at:
          type: string
          format: date-time
          example: 2026-04-20T10:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-04-20T12:00:00Z
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTHENTICATION_FAILED
            message:
              type: string
              example: Invalid credentials
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The given data was invalid
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: The email field is required
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint